requirements.txt
Django>=4.2,<5.0
djangorestframework
Pillow
django-environ
psycopg2-binary